+++ src/app/sensor-dialog/sensor-dialog.component.ts
@@ -81,6 +81,10 @@ export class SensorDialogComponent implements OnInit {
         else if(instrument.status === "Lost"){
           self.lostSensors ++;
         }
+
+        if(instrument.vdatum === undefined){
+          self.setVDatum(instrument);
+        }
       });
     }
 
@@ -126,6 +130,14 @@ export class SensorDialogComponent implements OnInit {
         }
       })
     })
+  }
+
+  setVDatum(instrument) {
+    this.siteService
+    .getVDatum(instrument.vdatum_id)
+    .subscribe((results) => {
+      instrument.vdatum = results.datum_name;
+    });
   }
 
   sortSensorFilesData(sort: Sort) {