+++ CHANGELOG.md
@@ -4,7 +4,28 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
 ## [Unreleased](https://github.com/USGS-WiM/stnweb2/tree/dev)
+
+### Added
+
+ -  Clear 'X' button and progress spinner to event state and event type filters
+ -  Info tooltips to fields in edit/add modals where units are converted on submit
+ -  Active highlighting on hover, button click and row add to site page tables
+ -  When new row item is added and table page is full, jump to next page
+ -  Active highlighting on hover and details button click to map filter results table
+ 
+### Changed
+
+ -  Disabled sensor retrieve button when no event is selected
+ 
+ ### Fixed
+ 
+ -  Fields not populating in details modals
+ -  Unedited rows removed from reference datum table after edit
+ -  Update counts in site page tabs on edit/add/delete
+
+## [v0.8.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.8.0) - 2022-06-05
 
 ### Added
 
@@ -18,7 +39,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Alert for deploying and retrieving sensors when no event selected
  -  Added ability to create a peak
  -  Added tooltips to all add button icons
+ -  Added ability to delete a peak
  -  Added Description and Is Destroyed columns in reference datum table
+ -  Added ability to retrieve a sensor
  
 ### Changed
 
@@ -27,6 +50,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Improved responsiveness of tables in details modals
  -  Autopopulate source and agency when adding file in site edit form
  -  Disabled add button when no event selected for Peaks, Sensors, HWMs and files (except site files)
+ -  Disabled peak edit when no event selected
  -  Date Recovered to Date Last Checked in reference datum detail and edit modals
  
  ### Fixed
@@ -34,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  -  Time reset in sensor edit modal when Cancel Edits is clicked
  -  Time validation in sensor edit
  -  Inability to login after logging out
+ -  Duplicated disabled buttons when incorrect role and no event selected
 
 ## [v0.7.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.7.0) - 2022-01-17